Complete working code in C

+++ sw/include/adcsampling.h
@@ -6,6 +6,6 @@
 #include <read_hwvariation.h>
 
 // Declare function prototypes
-float processADCValue(uint8_t *adc_rawinput, hw_variation_enum_t hw_variation);
+float processADCValue(float *adc_rawinput, hw_variation_enum_t hw_variation);
 
 #endif // ADCSAMPLING_H
